Develop

+++ src/main/java/ru/springtest/service/implementation/CustomerAccountServiceImplementation.java
@@ -1,6 +1,9 @@
 package ru.springtest.service.implementation;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.CachePut;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.springtest.domain.Account;
@@ -22,6 +25,7 @@ public class CustomerAccountServiceImplementation implements CustomerAccountServ
     private final AccountRepository accountRepository;
     private final CustomerAccountMapper mapper;
 
+    @CachePut(value = "customer", key = "#result.customerId()")
     @Transactional
     @Override
     public CustomerAccountResponseDto createWithAccount(CustomerAccountCreateUpdateDto dto) {
@@ -30,19 +34,21 @@ public class CustomerAccountServiceImplementation implements CustomerAccountServ
         return mapper.toDto(customer, account);
     }
 
+    @Cacheable(value = "customer", key = "#id")
     @Transactional
     @Override
     public CustomerAccountResponseDto getById(UUID id) {
-        Customer customer = customerRepository.findById(id).orElseThrow(()->new NotFoundException("Customer not found with id:" + id));
+        Customer customer = customerRepository.findById(id).orElseThrow(()->new NotFoundException("Customer not found with id: " + id));
         return mapper.toDto(
                 customer,
                 accountRepository.findAccountsByCustomerId(id));
     }
 
+    @CachePut(value = "customer", key = "#result.customerId()")
     @Override
     @Transactional
     public CustomerAccountResponseDto updateWithAccount(CustomerAccountCreateUpdateDto dto, UUID id) {
-        Customer customer = customerRepository.findById(id).orElseThrow(()->new NotFoundException("Customer not found with id:" + id));
+        Customer customer = customerRepository.findById(id).orElseThrow(()->new NotFoundException("Customer not found with id: " + id));
         mapper.changeCustomer(customer, dto);
         customerRepository.save(customer);
         Account account = accountRepository.findAccountsByCustomerId(id);
@@ -51,9 +57,13 @@ public class CustomerAccountServiceImplementation implements CustomerAccountServ
         return mapper.toDto(customer, account);
     }
 
+    @CacheEvict(value = "customer", key = "#id")
     @Override
     @Transactional
     public void delete(UUID id) {
+        if (!customerRepository.existsById(id)) {
+            throw new NotFoundException("Customer not found with id: " + id);
+        }
         customerRepository.deleteById(id);
     }
